Turn knob to 11 on warnings causing build failures

+++ src/main/java/frc/robot/TeamTalonSRX.java
@@ -117,7 +117,7 @@ public class TeamTalonSRX extends WPI_TalonSRX {
       return 0;
     }
     double currentSpeed = getSelectedSensorVelocity();
-    return (double) (getClosedLoopTarget() - currentSpeed);
+    return (getClosedLoopTarget() - currentSpeed);
   }
 
   public void configureWithPidParameters(PidParameters pidParameters, int pidSlotIndex) {
